Scoring fixes, PDF Inter fonts, dashboard updates, agent docs, and gitignore hardening
- Fix engine scorer edge cases (health, improvement, persistence, MoS) with three new test suites
- Add Inter font family (Regular/Bold/Italic) for PDF rendering
- Refresh all dashboard templates and routes (pipeline, settings, stocks)
- Add docs/agents/ (domain, issue-tracker, triage-labels) and migrate issues/ to GitHub Issues
- Update CLAUDE.md files across modules to reflect current architecture
- Harden .gitignore: exclude .flask_secret and .claude/worktrees/

+++ db/db_maintenance.py
@@ -33,7 +33,7 @@ def clean_bad_dps(max_yield_non_reit: float = 20.0,
     # Join financials with latest price and REIT flag
     rows = conn.execute("""
         SELECT f.ticker, f.year, f.dps, p.close, s.is_reit,
-               round(f.dps / p.close * 100.0, 1) AS yield_pct
+               round((f.dps / p.close * 100.0)::numeric, 1) AS yield_pct
         FROM financials f
         JOIN stocks s ON f.ticker = s.ticker
         JOIN (
